Flatten error check, also move test to type inference tests which is a more sensible location.
